requirement.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
PyMySQL==0.10.1
pytz==2020.4
sqlparse==0.4.1
